API/TripCraft_AI/requirements.txt

fastapi
uvicorn
pydantic
python-multipart
httpx
asyncio
python-dotenv
loguru
qdrant-client
sentence-transformers
transformers
Pillow
crawl4ai
ollama
websockets
aiofiles
python-jose
passlib
bcrypt
openai-whisper # Free speech recognition
pandas
numpy
dotenv
pydantic-settings


icalendar==5.0.11
folium
beautifulsoup4
duckduckgo-search
tenacity
pytz